Airbyte-DBT-SQL/Cap12/requirements.txt
pandas
sqlalchemy
psycopg2-binary
langchain_core
langchain_community